backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.3.2
pymongo==4.6.3
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
PyJWT==2.9.0
httpx==0.27.2
anthropic==0.34.2
python-multipart==0.0.12
sqlalchemy==2.0.36
pandas==2.2.3
openpyxl==3.1.5
groq==0.11.0
python-dotenv==1.0.1
pandas
openpyxl
pdfplumber
premailer
razorpay
chromadb
sentence-transformers